+++ source/src/devel/domain_insertion/FusePosesNtoCMover.cc
@@ -292,14 +292,14 @@ FusePosesNtoCMover::setup_relax_fold_tree(
 
 		//sanity check: we need to make sure that we don't have edges
 		//originating or terminating in the fusion regions
-		if ( ((e.start() >= (int) fusion_regions_[1].first) && (e.start() <= (int) fusion_regions_[1].second))
-				|| ((e.stop() >= (int) fusion_regions_[1].first) && (e.stop() <= (int) fusion_regions_[1].second))
-				|| ((e.start() >= (int) fusion_regions_[2].first) && (e.start() <= (int) fusion_regions_[2].second))
-				|| ((e.stop() >= (int) fusion_regions_[2].first) && (e.stop() <= (int) fusion_regions_[2].second)) ) {
+		if ( ((e.start() >= fusion_regions_[1].first) && (e.start() <= fusion_regions_[1].second))
+				|| ((e.stop() >= fusion_regions_[1].first) && (e.stop() <= fusion_regions_[1].second))
+				|| ((e.start() >= fusion_regions_[2].first) && (e.start() <= fusion_regions_[2].second))
+				|| ((e.stop() >= fusion_regions_[2].first) && (e.stop() <= fusion_regions_[2].second)) ) {
 			utility_exit_with_message("Incipient fold tree has edges originating or terminating in the fusion regions, don't know what to do.");
 		}
 		//is this the chemical edge spanning the chain 1 fusion?
-		if ( (e.start() < (int) fusion_regions_[1].first) && (e.stop() > (int) fusion_regions_[1].second) && !e.is_jump() ) {
+		if ( (e.start() < fusion_regions_[1].first) && (e.stop() > fusion_regions_[1].second) && !e.is_jump() ) {
 			new_ft.add_edge( e.start(), fusion_regions_[1].first - 1, e.label() ); //a.
 			new_ft.add_edge( fusion_regions_[1].first -1, fuse_positions_[1], e.label() ); //b.
 			new_ft.add_edge( fusion_regions_[1].first-1, fuse_positions_[1] + 1, jump_count ); //c.
@@ -307,7 +307,7 @@ FusePosesNtoCMover::setup_relax_fold_tree(
 			++jump_count;
 			new_ft.add_edge( fuse_positions_[1] + 1, e.stop(), e.label() ); //d.
 			chain1_span_end = e.stop();
-		} else if ( (e.start() < (int) fusion_regions_[2].first) && (e.stop() > (int) fusion_regions_[2].second) && !e.is_jump() ) {
+		} else if ( (e.start() < fusion_regions_[2].first) && (e.stop() > fusion_regions_[2].second) && !e.is_jump() ) {
 			//or is this the chemical edge spanning the chain 2 fusion?
 			new_ft.add_edge( e.start(), fuse_positions_[2], e.label() ); //e.
 			if ( chain1_span_end ==0 ) utility_exit_with_message("Edge spanning a later chain was first detected in FoldTree.");
@@ -852,7 +852,7 @@ SetupCoiledCoilFoldTreeMover::apply( core::pose::Pose & pose )
 
 		std::cout << "looping edge with label " << e. label() << ", start " << e.start() << " and stop " << e.stop() << std::endl;
 
-		if ( (e.label() == core::kinematics::Edge::PEPTIDE) && (e.start() < (int) cut_seqpos) && (e.stop() > (int) cut_seqpos) ) {
+		if ( (e.label() == core::kinematics::Edge::PEPTIDE) && (e.start() < cut_seqpos) && (e.stop() > cut_seqpos) ) {
 			std::cout << "MEEP" << std::endl;
 			new_ft.add_edge( e.start(), cut_seqpos, e.label() );
 			new_ft.add_edge( e.stop(), cut_seqpos + 1, e.label() );
